old_project/requirements.txt
face_recognition==1.3.0
faiss_cpu==1.9.0
fer==22.5.1
mediapipe==0.10.15
mtcnn==1.0.0
numpy==1.26.4
opencv_python==4.10.0.84
pandas==2.2.3
tensorflow==2.18.0
pytato
faiss-cpu~=1.9.0
opencv-python~=4.10.0.84
tqdm~=4.67.1
torch~=2.2.2
pillow~=10.2.0
torchvision~=0.17.2
pyyaml~=6.0.2
matplotlib~=3.9.2
seaborn~=0.13.2
ipython~=8.30.0
requests~=2.32.3
psutil~=6.1.0
thop~=0.1.1-2209072238
pycocotools~=2.0.8
face-recognition~=1.3.0